requirements.txt

streamlit
streamlit-chat
langchain
langchain-experimental
openai
python-dotenv
elasticsearch
tiktoken
pypdf